requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
python-multipart==0.0.9
descope>=1.7.0
stripe>=5.0.0
asyncpg>=0.29.0
greenlet>=3.0.0  # Required for SQLAlchemy async
pg8000==1.30.5  # Pure Python PostgreSQL driver
apscheduler==3.10.4
pytz==2024.1
tzlocal==5.2
pydantic==2.6.1
pydantic-settings==2.1.0
pytest==8.3.5
pytest-asyncio==1.2.0
boto3==1.34.131
redis[hiredis]>=5.0.0
pusher==3.3.0
httpx==0.25.0
bleach>=6.0.0